Increase number of shapes supported by protobuf
`Path2D` is now transferable in addition to to `Area`, with the latter now being implemented on top fo the former.

All `Ellipse2D` are now transferable, whereas previously only `Ellipse2D.Float` was allowed. We always use double
precision for ellipses.

+++ src/main/java/net/rptools/maptool/client/tool/drawing/OvalTool.java
@@ -66,7 +66,7 @@ public class OvalTool extends AbstractDrawingTool implements MouseMotionListener
       paintTransformed(
           g,
           renderer,
-          new ShapeDrawable(new Ellipse2D.Float(oval.x, oval.y, oval.width, oval.height)),
+          new ShapeDrawable(new Ellipse2D.Double(oval.x, oval.y, oval.width, oval.height)),
           pen);
 
       ToolHelper.drawBoxedMeasurement(
@@ -101,7 +101,7 @@ public class OvalTool extends AbstractDrawingTool implements MouseMotionListener
 
         completeDrawable(
             getPen(),
-            new ShapeDrawable(new Ellipse2D.Float(oval.x, oval.y, oval.width, oval.height), true));
+            new ShapeDrawable(new Ellipse2D.Double(oval.x, oval.y, oval.width, oval.height), true));
         oval = null;
       }
